use en and em dashes
For grammatical and aesthetic correctness, we propose that the delimiter between the time start and time end of each period be a spaced en dash, rather than a spaced hyphen. Furthermore, we propose that the delimiter between the "start/end" string and the period duration be a spaced em dash.