add error on events loaded message
sometimes, our internet connection is dysfunctional for reasons outside our control. In this case, WATT used to describe the situation as having no events to display, whereas in reality, the events failed to be fetched. In this commit, I propose a solution: a variable eventsMessage passed down from the App class into the Events class, which takes on one of three values (the empty string, signifying a successful load; a "loading string"; and an "error string"). This solution may be extended to include a refresh button, as suggested by Kevin.