Update client side alternate JSON
Consider having just the raw JSON in client instead of wrapping in a `ts` file (CRA comes with `json-loader` already installed so importing from JSON is not a hassle). The benefits of wrapping are that you can type check directly in the file, but type checking can also be done at the component level by importers of the JSON. As it stands, the wrapping of the JSON disproportionately affects the repository language %s and is also not really idiomatic, though if the JSONs are ever depended on by multiple components or files then type checking individually in each one can be a bit of a pain.