Update client dependencies
I experimented with using `Dialog.Panel` as recommended in https://github.com/tailwindlabs/headlessui/releases/tag/%40headlessui%2Freact%40v1.6.0, but it has undesired behavior when combined with nested modals (ie. staff -> club). we could consider trying to use this again in the future. Also, updating to `headlessui@1.6.0` caused the close animations of CenteredModal not to render which was certainly odd.